# auth-client/auth-java-client/swagger-generator/v1.0.3/swagger.yaml
swagger: '2.0'
info:
  description: 更多信息请联系基础框架团队
  version: 1.0.3
  title: PAAS平台中央认证和授权系统
  termsOfService: 'http://localhost/'
host: localhost
basePath: /
tags:
  - name: o-auth-2-endpoint
    description: O Auth 2 Endpoint
paths:
  /oauth2/authorize:
    post:
      tags:
        - o-auth-2-endpoint
      summary: OAuth2授权点
      description: 授权点颁发authorization code
      operationId: authorizeUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: clientVO
          description: clientVO
          required: true
          schema:
            $ref: '#/definitions/ClientVO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AuthCodeVO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /oauth2/introspect:
    post:
      tags:
        - o-auth-2-endpoint
      summary: OAuth2令牌检查点
      description: 检查令牌
      operationId: introspectTokenUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: token
          in: query
          description: token
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ValidityVO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /oauth2/revoke:
    post:
      tags:
        - o-auth-2-endpoint
      summary: OAuth2令牌吊销点
      description: 吊销令牌
      operationId: revokeTokenUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: token
          in: query
          description: token
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: boolean
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /oauth2/token:
    post:
      tags:
        - o-auth-2-endpoint
      summary: OAuth2令牌颁发点
      description: 颁发令牌access/refresh token
      operationId: issueTokenUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: grant_type
          in: query
          description: grant_type
          required: true
          type: string
        - name: code
          in: query
          description: code
          required: false
          type: string
        - name: redirect_uri
          in: query
          description: redirect_uri
          required: false
          type: string
        - name: refresh_token
          in: query
          description: refresh_token
          required: false
          type: string
        - name: username
          in: query
          description: username
          required: false
          type: string
        - name: password
          in: query
          description: password
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/OAuth2AccessToken'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  AuthCodeVO:
    type: object
    properties:
      clientName:
        type: string
      code:
        type: string
      expiration:
        type: string
        format: date-time
      id:
        type: integer
        format: int64
      insertTime:
        type: string
        format: date-time
      redirectUrl:
        type: string
      userName:
        type: string
  ClientVO:
    type: object
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      description:
        type: string
      id:
        type: integer
        format: int64
      ownerName:
        type: string
      redirectUrl:
        type: string
      rememberChoice:
        type: string
      scopes:
        type: array
        items:
          type: string
  OAuth2AccessToken:
    type: object
    properties:
      access_token:
        type: string
      token_type:
        type: string
      refresh_token:
        type: string
      expires_in:
        type: integer
        format: int32
  ValidityVO:
    type: object
    properties:
      isValid:
        type: boolean